requirements.txt
Flask==0.12.2
SQLAlchemy==1.1.10
python-dotenv==0.6.4
requests==2.14.2
zappa>=0.17.6
Flask-GraphQL==1.4.1
graphene-sqlalchemy==1.1.1
psycopg2
coinbase